Added camera movement

+++ Graphics/src/Engine.h
@@ -15,7 +15,7 @@ public:
 
 private:
 	int start();
-	int update(float deltaTime);
+	int update(double deltaTime);
 	int draw();
 	int end();
